+++ CHANGELOG.md
@@ -7,4 +7,10 @@
 - Allow selecting specific fields from hash.get and hash.search
 
 ## 0.3.0 - 2016-09-04
-- Allow removing specified hash by key 
+- Allow removing specified hash by key
+
+## 0.4.0 - 2016-09-04
+- Add timestamps(createdAt & updatedAt) to the saved object
+- Ensure update & save update timestamps
+- Allow multisave
+- Improve code JSDOCS